requirements.txt
fastapi
uvicorn
python-multipart
jinja2
openai-whisper
torch
requests